Move index.js into legacy/index.js
